requirements.txt
slack_bolt
schedule
requests
PyYAML
transformers
torch
cachetools
python-dateutil
